Updated Code: 1. Added Update Property 2. Added Delete Property 3. Updated Update Method 4. Updated Delete Method 5. Updated Model Mapper

+++ src/main/java/org/generationcp/bms/ontology/services/impl/OntologyModelServiceImpl.java
@@ -16,6 +16,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.ArrayList;
 
@@ -45,8 +46,13 @@ public class OntologyModelServiceImpl implements OntologyModelService {
     public MethodResponse getMethod(Integer id) throws MiddlewareQueryException {
         Method method = ontologyService.getMethod(id);
         if(method == null) return null;
+        boolean deletable = true;
+        if(ontologyService.isTermReferred(id)) deletable = false;
         ModelMapper mapper = OntologyMapper.methodMapper();
-        return mapper.map(method, MethodResponse.class);
+        MethodResponse response = mapper.map(method, MethodResponse.class);
+        if(deletable) response.setEditableFields(new ArrayList<>(Arrays.asList("description")));
+        response.setDeletable(deletable);
+        return response;
     }
 
     @Override
@@ -56,14 +62,19 @@ public class OntologyModelServiceImpl implements OntologyModelService {
     }
 
     @Override
-    public void updateMethod(Integer id, MethodRequest request) throws MiddlewareQueryException, MiddlewareException {
+    public boolean updateMethod(Integer id, MethodRequest request) throws MiddlewareQueryException, MiddlewareException {
+        if(ontologyService.isTermReferred(id)) return false;
         Method method = new Method(new Term(id, request.getName(), request.getDescription()));
         ontologyService.updateMethod(method);
+        return true;
     }
 
     @Override
-    public void deleteMethod(Integer id) throws MiddlewareQueryException {
+    public boolean deleteMethod(Integer id) throws MiddlewareQueryException {
+        boolean isReferred = ontologyService.isTermReferred(id);
+        if(isReferred) return false;
         ontologyService.deleteMethod(id);
+        return true;
     }
 
     @Override
@@ -84,12 +95,17 @@ public class OntologyModelServiceImpl implements OntologyModelService {
     public PropertyResponse getProperty(Integer id) throws MiddlewareQueryException {
         Property property = ontologyService.getPropertyById(id);
         if (property == null) return null;
+        boolean deletable = true;
+        if(ontologyService.isTermReferred(id)) deletable = false;
         ModelMapper mapper = OntologyMapper.propertyMapper();
-        return mapper.map(property, PropertyResponse.class);
+        PropertyResponse response = mapper.map(property, PropertyResponse.class);
+        if(deletable) response.setEditableFields(new ArrayList<>(Arrays.asList("description")));
+        response.setDeletable(deletable);
+        return response;
     }
 
     @Override
-    public GenericResponse addProperty(PropertyRequest request) throws MiddlewareQueryException {
+    public GenericResponse addProperty(PropertyRequest request) throws MiddlewareQueryException, MiddlewareException {
         return new GenericResponse(ontologyService.addProperty(request.getName(), request.getDescription(), request.getCropOntologyId(), request.getClasses()).getId());
     }
 
@@ -106,6 +122,21 @@ public class OntologyModelServiceImpl implements OntologyModelService {
             properties.add(propertyDTO);
         }
         return properties;
+    }
+
+    @Override
+    public boolean deleteProperty(Integer id) throws MiddlewareQueryException, MiddlewareException {
+        boolean isReferred = ontologyService.isTermReferred(id);
+        if(isReferred) return false;
+        ontologyService.deleteProperty(id);
+        return true;
+    }
+
+    @Override
+    public boolean updateProperty(Integer id, PropertyRequest request) throws MiddlewareQueryException, MiddlewareException {
+        if(ontologyService.isTermReferred(id)) return false;
+        ontologyService.updateProperty(id, request.getName(), request.getDescription(), request.getCropOntologyId(), request.getClasses());
+        return true;
     }
 
     @Override
